# public/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
        "/api/v1/crewMembers": {
            "get": {
                "tags": [
                    "Crew Members"
                ],
                "summary": "Get a list of crew members",
                "operationId": "2cae5a7f13689922c000a3fc7c990c9b",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "role",
                        "in": "query",
                        "description": "The role of the crew member",
                        "required": false,
                        "schema": {
                            "$ref": "#/components/schemas/CrewMemberRole"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of crew members",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CrewMember"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Out of range",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Out of range"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Bad request"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Crew Members"
                ],
                "summary": "Add a new crew member",
                "operationId": "4966672b2daafa0f0596d6c275cb3cf1",
                "requestBody": {
                    "description": "Crew member data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CrewMember"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Crew member created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully created Crew Member"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid input"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Only admins can add crew members",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Only admins can add crew members"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "208": {
                        "description": "Crew Member already exists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Crew Member already exists"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/movies": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Get a page of movies",
                "operationId": "673717dae9ef9ddea1ed552297a693a8",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Page of movies retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "results": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Movie"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "total_pages": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Page not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Page not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Movies"
                ],
                "summary": "Create a new movie",
                "operationId": "1bcee0cc618464780d2c23d7b3367600",
                "requestBody": {
                    "description": "Movie data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Movie"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Movie created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully added Movie"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid input"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Only admins can create movies",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Only admins can create movies"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Movie already exists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Movie already exists"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CrewMember": {
                "title": "CrewMember model",
                "description": "CrewMember model",
                "properties": {
                    "id": {
                        "description": "ID of the crew member",
                        "type": "integer",
                        "format": "int64",
                        "example": 1,
                        "nullable": true
                    },
                    "name": {
                        "description": "Name of the crew member",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "photo": {
                        "description": "Photo of the crew member",
                        "type": "string",
                        "example": "path/to/image.jpg"
                    },
                    "role": {
                        "description": "Role of the crew member",
                        "type": "string",
                        "example": "Director"
                    }
                },
                "type": "object"
            },
            "Movie": {
                "title": "Movie model",
                "description": "Movie model",
                "properties": {
                    "id": {
                        "description": "ID of the movie",
                        "type": "integer",
                        "format": "int64",
                        "example": 1,
                        "nullable": true
                    },
                    "title": {
                        "description": "Title of the movie",
                        "type": "string",
                        "example": "The Godfather"
                    },
                    "running_time": {
                        "description": "Running time of the movie",
                        "type": "integer",
                        "example": 175
                    },
                    "cover_photo": {
                        "description": "Cover photo of the movie",
                        "type": "string",
                        "example": "path/to/image.jpg"
                    },
                    "overview": {
                        "description": "Overview of the movie",
                        "type": "string",
                        "example": "The aging patriarch of an organized crime dynasty transfers control of his clandestine empire to his reluctant son."
                    },
                    "release_date": {
                        "description": "Release date of the movie",
                        "type": "string",
                        "example": "1972-03-14T00:00:00+00:00"
                    }
                },
                "type": "object"
            },
...
            "CrewMemberRole": {
                "title": "CrewMemberRole",
                "description": "CrewMemberRole enum",
                "properties": {
                    "Actor": {
                        "description": "Represents an actor",
                        "type": "object",
                        "enum": [
                            "actor"
                        ]
                    },
                    "Director": {
                        "description": "Represents a director",
                        "type": "object",
                        "enum": [
                            "director"
                        ]
                    }
                },
                "type": "string",
                "enum": [
                    "actor",
                    "director"
                ],
                "example": "actor"
            }
        }
    }
}